# docs/api-spec.yaml
openapi: 3.1.0
info:
  title: DEVLINK API
  description: API service for the Distributed Link Intelligence Platform.
  version: 0.1.0
paths:
...
  /api/v1/workspaces:
    get:
      tags:
      - workspaces
      summary: List workspaces
      description: List all workspaces where the authenticated user is a member.
      operationId: list_workspaces_api_v1_workspaces_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/WorkspaceResponse'
                type: array
                title: Response List Workspaces Api V1 Workspaces Get
        '401':
          description: Authentication is required.
        '403':
          description: Required workspace role is missing.
        '404':
          description: Workspace or member not found.
        '409':
          description: Workspace membership already exists.
        '422':
          description: Request validation failed.
      security:
      - OAuth2PasswordBearer: []
    post:
      tags:
      - workspaces
      summary: Create a workspace
      description: Create a workspace and make the authenticated user its owner.
      operationId: create_workspace_api_v1_workspaces_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkspaceCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceResponse'
        '401':
          description: Authentication is required.
        '403':
          description: Required workspace role is missing.
        '404':
          description: Workspace or member not found.
        '409':
          description: Workspace membership already exists.
        '422':
          description: Request validation failed.
      security:
      - OAuth2PasswordBearer: []
  /api/v1/workspaces/{workspace_id}:
    get:
      tags:
      - workspaces
      summary: Get a workspace
      description: Return workspace metadata for users with at least viewer access.
      operationId: get_workspace_api_v1_workspaces__workspace_id__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: workspace_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceResponse'
        '401':
          description: Authentication is required.
        '403':
          description: Required workspace role is missing.
        '404':
          description: Workspace or member not found.
        '409':
          description: Workspace membership already exists.
        '422':
          description: Request validation failed.
    patch:
      tags:
      - workspaces
      summary: Update a workspace
      description: Update workspace metadata. Requires admin or owner role.
      operationId: update_workspace_api_v1_workspaces__workspace_id__patch
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: workspace_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkspaceUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceResponse'
        '401':
          description: Authentication is required.
        '403':
          description: Required workspace role is missing.
        '404':
          description: Workspace or member not found.
        '409':
          description: Workspace membership already exists.
        '422':
          description: Request validation failed.
    delete:
      tags:
      - workspaces
      summary: Delete a workspace
      description: Delete a workspace. Requires owner role.
      operationId: delete_workspace_api_v1_workspaces__workspace_id__delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: workspace_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      responses:
        '204':
          description: Successful Response
        '401':
          description: Authentication is required.
        '403':
          description: Required workspace role is missing.
        '404':
          description: Workspace or member not found.
        '409':
          description: Workspace membership already exists.
        '422':
          description: Request validation failed.
  /api/v1/workspaces/{workspace_id}/members:
    post:
      tags:
      - workspaces
      summary: Add a workspace member
      description: Add a user to a workspace. Requires admin or owner role.
      operationId: add_member_api_v1_workspaces__workspace_id__members_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: workspace_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkspaceMemberCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceMemberResponse'
        '401':
          description: Authentication is required.
        '403':
          description: Required workspace role is missing.
        '404':
          description: Workspace or member not found.
        '409':
          description: Workspace membership already exists.
        '422':
          description: Request validation failed.
    get:
      tags:
      - workspaces
      summary: List workspace members
      description: List workspace members. Requires at least viewer role.
      operationId: list_members_api_v1_workspaces__workspace_id__members_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: workspace_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkspaceMemberResponse'
                title: Response List Members Api V1 Workspaces  Workspace Id  Members
                  Get
        '401':
          description: Authentication is required.
        '403':
          description: Required workspace role is missing.
        '404':
          description: Workspace or member not found.
        '409':
          description: Workspace membership already exists.
        '422':
          description: Request validation failed.
  /api/v1/workspaces/{workspace_id}/members/{user_id}:
    patch:
      tags:
      - workspaces
      summary: Update a workspace member
      description: Change a member role. Requires admin or owner role.
      operationId: update_member_api_v1_workspaces__workspace_id__members__user_id__patch
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: workspace_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkspaceMemberUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceMemberResponse'
        '401':
          description: Authentication is required.
        '403':
          description: Required workspace role is missing.
        '404':
          description: Workspace or member not found.
        '409':
          description: Workspace membership already exists.
        '422':
          description: Request validation failed.
    delete:
      tags:
      - workspaces
      summary: Remove a workspace member
      description: Remove a member from a workspace. Requires admin or owner role.
      operationId: remove_member_api_v1_workspaces__workspace_id__members__user_id__delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: workspace_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
      responses:
        '204':
          description: Successful Response
        '401':
          description: Authentication is required.
        '403':
          description: Required workspace role is missing.
        '404':
          description: Workspace or member not found.
        '409':
          description: Workspace membership already exists.
        '422':
          description: Request validation failed.
...
components:
  schemas:
...
    WorkspaceCreate:
      properties:
        name:
          type: string
          maxLength: 120
          minLength: 1
          title: Name
      additionalProperties: false
      type: object
      required:
      - name
      title: WorkspaceCreate
      description: Request body for creating a workspace.
      example:
        name: Engineering
    WorkspaceMemberCreate:
      properties:
        user_id:
          type: string
          format: uuid
          title: User Id
        role:
          type: string
          title: Role
          default: viewer
      additionalProperties: false
      type: object
      required:
      - user_id
      title: WorkspaceMemberCreate
      description: Request body for adding a workspace member.
      example:
        role: editor
        user_id: 29e61574-f3ae-4f08-8ba4-c770659d0d6e
    WorkspaceMemberResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        workspace_id:
          type: string
          format: uuid
          title: Workspace Id
        user_id:
          type: string
          format: uuid
          title: User Id
        role:
          type: string
          title: Role
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - workspace_id
      - user_id
      - role
      - created_at
      title: WorkspaceMemberResponse
      description: Workspace member response payload.
      example:
        created_at: '2026-06-15T12:00:00Z'
        id: f88a2925-71e2-4144-9450-44af3a7d12ab
        role: editor
        user_id: 29e61574-f3ae-4f08-8ba4-c770659d0d6e
        workspace_id: b9300502-d860-4d26-ad73-4f88724f9fb0
    WorkspaceMemberUpdate:
      properties:
        role:
          type: string
          title: Role
      additionalProperties: false
      type: object
      required:
      - role
      title: WorkspaceMemberUpdate
      description: Request body for updating a workspace member role.
      example:
        role: viewer
    WorkspaceResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        name:
          type: string
          title: Name
        owner_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Owner Id
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - name
      - owner_id
      - created_at
      - updated_at
      title: WorkspaceResponse
      description: Workspace response payload.
      example:
        created_at: '2026-06-15T12:00:00Z'
        id: b9300502-d860-4d26-ad73-4f88724f9fb0
        name: Engineering
        owner_id: d7b5f7a2-8a6c-4a46-9443-6480a4f7aa1e
        updated_at: '2026-06-15T12:00:00Z'
    WorkspaceUpdate:
      properties:
        name:
          type: string
          maxLength: 120
          minLength: 1
          title: Name
      additionalProperties: false
      type: object
      required:
      - name
      title: WorkspaceUpdate
      description: Request body for updating workspace metadata.
      example:
        name: Platform Engineering
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/auth/login